vcf variant analyser
script to analyse  vcf files in dir and subdir. Outputs three figures with data from all files selected. First plot is simple type of variation occuring in %. Second plot is heatmap of sequence of 5 bases with mutation and last figure is also an heatmap with only snps (similarly as first figure). This script should be competible with the HPC